Move bitcoin transaction in module
